refactor(opensearch-migration): consolidate startup config banner in IndexStartupValidator
Move IndexStartupValidator into com.dotcms.content.index.opensearch so the
config summary and the connection outcome are logged together as one startup
story under one logger — instead of the config banner firing from the provider
at CDI-construction time (decoupled from the validation moment) and the outcome
banner firing later from the validator.

- ConfigurableOpenSearchProvider.configFromProperties() is now package-private
  static; the validator re-resolves the exact same config for the banner. No
  config accessor is exposed outside the package and the OSClientProvider CDI
  contract is untouched.
- The provider no longer logs the INFO banner; it keeps a quiet DEBUG line so
  runtime (re)builds stay observable without duplicating the banner.
- Update the import at the single call site (ContentletIndexAPIImpl).

No changes to existing Elasticsearch classes.
